+++ CHANGELOG.md
@@ -12,6 +12,10 @@
 * Three inner (private) attributes (`formName`, `selectBoxName`, and `fieldNamePrefix`) have
 been made public.
 
+### `Cape.ResourceAgent`
+
+* Add a new class called `Cape.Resource`, which performs CRUD operation using `fetch` API.
+
 ## 1.1.0 (June, 7, 2015)
 
 ### `Cape.Component`