+++ CHANGELOG.md
@@ -23,7 +23,8 @@ been made public.
 
 ### `Cape.ResourceAgent`
 
-* Add a new class called `Cape.ResourceAgent`, which performs CRUD operation using `fetch` API.
+* Add a new class called `Cape.ResourceAgent`, which performs CRUD operations
+using `fetch` API.
 
 ## 1.1.0 (June, 7, 2015)