+++ CHANGELOG.md
@@ -24,7 +24,7 @@ been made public.
 
 ### `Cape.ResourceAgent`
 
-* Add a new class called `Cape.ResourceAgent`, which performs CRUD operations
+* Add a new class called `Cape.ResourceAgent`, which performs Ajax requests
 using `fetch` API.
 
 ## 1.1.0 (June, 7, 2015)